code/qpptk/requirements.txt
lmdb==1.2.1
lxml==4.6.4
matplotlib==3.4.3
msgpack_python==0.5.6
numpy==1.21.4
pandas==1.3.4
protobuf==3.19.1
scikit_learn==1.0.1
scipy==1.7.2
seaborn==0.11.2
setuptools==58.5.3
statsmodels==0.13.0
syct==0.4.3
toml==0.10.2
tqdm==4.62.3
